# PERF: MultiIndex.isin