- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.0.0.rst` file if fixing a bug or adding a new feature.

Improves perf for `MultiIndex.isin`.

```
       before           after         ratio
     [d47e0523]       [edd67b0a]
     <main>           <mi-isin> 
-     7.98±0.06ms      6.06±0.07ms     0.76  multiindex_object.Isin.time_isin_small('string')
-     5.70±0.07ms      4.02±0.03ms     0.71  multiindex_object.Isin.time_isin_small('int')
-     6.18±0.09ms      4.31±0.03ms     0.70  multiindex_object.Isin.time_isin_small('datetime')
-      42.7±0.5ms       6.73±0.2ms     0.16  multiindex_object.Isin.time_isin_large('string')
-      41.7±0.8ms       4.86±0.1ms     0.12  multiindex_object.Isin.time_isin_large('int')
-         136±2ms      4.87±0.07ms     0.04  multiindex_object.Isin.time_isin_large('datetime')
```